requirements.txt
flask==3.0.3
Werkzeug==3.0.3
stripe==9.12.0
python-dotenv==1.0.1
requests==2.32.3
anthropic>=0.40.0
sentence-transformers==3.0.1
numpy==1.26.4
pytest==8.3.3
pytest-cov==5.0.0
pyyaml==6.0.2
